Modify unit tests to use the new API
